requirements.txt
# --- Web API ---
fastapi
uvicorn
python-multipart
python-dotenv

# --- AI & ML Core ---
ultralytics
torch
torchvision

# --- Logic & Data Processing ---
opencv-python-headless
pillow
numpy
pandas
editdistance
requests
tqdm
scikit-learn
pydantic
